Fix broken link in readme file
