+++ native/src/main/java/io/ballerina/persist/cmd/PersistCmd.java
@@ -19,7 +19,6 @@ package io.ballerina.persist.cmd;
 
 import io.ballerina.cli.BLauncherCmd;
 import io.ballerina.persist.PersistToolsConstants;
-import io.ballerina.projects.ProjectEnvironmentBuilder;
 import picocli.CommandLine;
 
 import java.io.PrintStream;
@@ -39,8 +38,6 @@ import static io.ballerina.persist.PersistToolsConstants.COMPONENT_IDENTIFIER;
 public class PersistCmd implements BLauncherCmd {
 
     private static final PrintStream errStream = System.err;
-    public String sourcePath = "";
-    public ProjectEnvironmentBuilder projectEnvironmentBuilder;
 
     @Override
     public void execute() {
@@ -65,13 +62,5 @@ public class PersistCmd implements BLauncherCmd {
     public void printUsage(StringBuilder stringBuilder) {
         stringBuilder.append("  ballerina " + PersistToolsConstants.COMPONENT_IDENTIFIER).
                 append(System.lineSeparator());
-    }
-
-    public void setSourcePath(String sourceDir) {
-        this.sourcePath = sourceDir;
-    }
-
-    public void setEnvironmentBuilder(ProjectEnvironmentBuilder projectEnvironmentBuilder) {
-        this.projectEnvironmentBuilder = projectEnvironmentBuilder;
     }
 }
